requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pillow==10.4.0
pytesseract==0.3.10
language-tool-python==2.7.1
